ingestion/requirements.txt
yfinance>=0.2.40
pandas>=2.0
pyarrow>=14.0
google-cloud-storage>=2.14
google-cloud-bigquery>=3.17
db-dtypes>=1.2
jupyterlab>=4.0
